split up the settings ui
