+++ tsconfig.json
@@ -13,7 +13,8 @@
     "isolatedModules": true,
     "strictNullChecks": true,
     "jsx": "react",
-    "lib": ["DOM", "ES5", "ES6", "ES7"]
+    "lib": ["DOM", "ES5", "ES6", "ES7"],
+    "skipLibCheck": true
   },
   "include": [
     "src/**/*.ts",
